+++ operator/internal/manifests/ruler.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 	"path"
 
+	"github.com/ViaQ/logerr/v2/kverrors"
 	"github.com/grafana/loki/operator/internal/manifests/internal/config"
+	"github.com/imdario/mergo"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -20,6 +22,12 @@ func BuildRuler(opts Options) ([]client.Object, error) {
 	statefulSet := NewRulerStatefulSet(opts)
 	if opts.Flags.EnableTLSServiceMonitorConfig {
 		if err := configureRulerServiceMonitorPKI(statefulSet, opts.Name); err != nil {
+			return nil, err
+		}
+	}
+
+	if opts.Flags.EnableTLSGRPCServices {
+		if err := configureRulerGRPCServicePKI(statefulSet, opts.Name); err != nil {
 			return nil, err
 		}
 	}
@@ -201,7 +209,8 @@ func NewRulerStatefulSet(opts Options) *appsv1.StatefulSet {
 
 // NewRulerGRPCService creates a k8s service for the ruler GRPC endpoint
 func NewRulerGRPCService(opts Options) *corev1.Service {
-	l := ComponentLabels(LabelRulerComponent, opts.Name)
+	serviceName := serviceNameRulerGRPC(opts.Name)
+	labels := ComponentLabels(LabelRulerComponent, opts.Name)
 
 	return &corev1.Service{
 		TypeMeta: metav1.TypeMeta{
@@ -209,8 +218,9 @@ func NewRulerGRPCService(opts Options) *corev1.Service {
 			APIVersion: corev1.SchemeGroupVersion.String(),
 		},
 		ObjectMeta: metav1.ObjectMeta{
-			Name:   serviceNameRulerGRPC(opts.Name),
-			Labels: l,
+			Name:        serviceName,
+			Labels:      labels,
+			Annotations: serviceAnnotations(serviceName, opts.Flags.EnableCertificateSigningService),
 		},
 		Spec: corev1.ServiceSpec{
 			ClusterIP: "None",
@@ -222,7 +232,7 @@ func NewRulerGRPCService(opts Options) *corev1.Service {
 					TargetPort: intstr.IntOrString{IntVal: grpcPort},
 				},
 			},
-			Selector: l,
+			Selector: labels,
 		},
 	}
 }
@@ -230,8 +240,7 @@ func NewRulerGRPCService(opts Options) *corev1.Service {
 // NewRulerHTTPService creates a k8s service for the ruler HTTP endpoint
 func NewRulerHTTPService(opts Options) *corev1.Service {
 	serviceName := serviceNameRulerHTTP(opts.Name)
-	l := ComponentLabels(LabelRulerComponent, opts.Name)
-	a := serviceAnnotations(serviceName, opts.Flags.EnableCertificateSigningService)
+	labels := ComponentLabels(LabelRulerComponent, opts.Name)
 
 	return &corev1.Service{
 		TypeMeta: metav1.TypeMeta{
@@ -240,8 +249,8 @@ func NewRulerHTTPService(opts Options) *corev1.Service {
 		},
 		ObjectMeta: metav1.ObjectMeta{
 			Name:        serviceName,
-			Labels:      l,
-			Annotations: a,
+			Labels:      labels,
+			Annotations: serviceAnnotations(serviceName, opts.Flags.EnableCertificateSigningService),
 		},
 		Spec: corev1.ServiceSpec{
 			Ports: []corev1.ServicePort{
@@ -252,7 +261,7 @@ func NewRulerHTTPService(opts Options) *corev1.Service {
 					TargetPort: intstr.IntOrString{IntVal: httpPort},
 				},
 			},
-			Selector: l,
+			Selector: labels,
 		},
 	}
 }
@@ -260,6 +269,50 @@ func NewRulerHTTPService(opts Options) *corev1.Service {
 func configureRulerServiceMonitorPKI(statefulSet *appsv1.StatefulSet, stackName string) error {
 	serviceName := serviceNameRulerHTTP(stackName)
 	return configureServiceMonitorPKI(&statefulSet.Spec.Template.Spec, serviceName)
+}
+
+func configureRulerGRPCServicePKI(sts *appsv1.StatefulSet, stackName string) error {
+	caBundleName := signingCABundleName(stackName)
+	secretVolumeSpec := corev1.PodSpec{
+		Volumes: []corev1.Volume{
+			{
+				Name: caBundleName,
+				VolumeSource: corev1.VolumeSource{
+					ConfigMap: &corev1.ConfigMapVolumeSource{
+						LocalObjectReference: corev1.LocalObjectReference{
+							Name: caBundleName,
+						},
+					},
+				},
+			},
+		},
+	}
+
+	secretContainerSpec := corev1.Container{
+		VolumeMounts: []corev1.VolumeMount{
+			{
+				Name:      caBundleName,
+				ReadOnly:  false,
+				MountPath: caBundleDir,
+			},
+		},
+		Args: []string{
+			// Enable GRPC over TLS for ruler client
+			"-ruler.client.tls-enabled=true",
+			fmt.Sprintf("-ruler.client.tls-ca-path=%s", signingCAPath()),
+		},
+	}
+
+	if err := mergo.Merge(&sts.Spec.Template.Spec, secretVolumeSpec, mergo.WithAppendSlice); err != nil {
+		return kverrors.Wrap(err, "failed to merge volumes")
+	}
+
+	if err := mergo.Merge(&sts.Spec.Template.Spec.Containers[0], secretContainerSpec, mergo.WithAppendSlice); err != nil {
+		return kverrors.Wrap(err, "failed to merge container")
+	}
+
+	serviceName := serviceNameRulerGRPC(stackName)
+	return configureGRPCServicePKI(&sts.Spec.Template.Spec, serviceName)
 }
 
 func ruleVolumeItems(tenants map[string]TenantConfig) []corev1.KeyToPath {