+++ pkg/querier/multi_tenant_querier.go
@@ -272,7 +272,7 @@ func (r relabel) relabel(original string) string {
 	}
 
 	lbls, _ = syntax.ParseLabels(original)
-	builder := labels.NewBuilder(lbls.WithoutLabels(defaultTenantLabel))
+	builder := labels.NewBuilder(lbls).Del(defaultTenantLabel)
 
 	// Prefix label if it conflicts with the tenant label.
 	if lbls.Has(defaultTenantLabel) {